requirements.txt
aiogram==3.13.1
pymongo==4.10.1
Pillow==10.4.0
imageio==2.36.0
imageio-ffmpeg==0.5.1